Some rennaming + separate ro and rw equality
