+++ src/base_json_equality.c
@@ -1,4 +1,4 @@
-#include "base_equality.h"
+#include "base_json_equality.h"
 
 /*******************************************************************************
 **                                 FUNCTIONS                                  **
