+++ src/rw_json_storage.h
@@ -99,6 +99,8 @@ typedef struct
 /*******************************************************************************
 **                                 FUNCTIONS                                  **
 *******************************************************************************/
+rw_array_t *init_rw_array();
+rw_dict_t *init_rw_dict();
 rw_json_t *init_rw_json(char is_array, rw_array_t *a, rw_dict_t *d);
 
 void rw_array_add_str(rw_array_t *a, string_t value);
